car_provider/requirements.txt
Django==3.2.15
djangorestframework==3.13.1
psycopg2-binary==2.8.6
django-filter==22.1
gunicorn==20.1.0
drf_yasg